feat: make an MFE accept external deps and use global exports
This MFE now doesn’t bundle react, react-dom, and parts of frontend-platform, relying on a parent to have them.  It has a “ChildPage” that attaches itself to global when it loads, which the parent can then pick up and render.

The only thing not really working is hot module replacement for this code when being rendered by a parent; a full page refresh happens instead.  This is because I couldn’t figure out how to get two HMR instances working at the same time.